Rename core.py to signal.py
