+++ magia/comb_select.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass
 from string import Template
 
 from .constant import Constant
-from .core import Signal
 from .data_struct import OPType
 from .factory import constant_like
+from .signal import Signal
 
 IF_ELSE_TEMPLATE = Template(
     "always_comb\n" \
