+++ magia/data_struct.py
@@ -12,7 +12,7 @@ from enum import Enum, IntEnum, auto
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from .core import Signal
+    from .signal import Signal
 
 
 class SignalDict(UserDict):
