+++ magia/signal.py
